Remove BaseRoute class
